+++ src/icoforge/gui/widgets/settings_panel.py
@@ -11,7 +11,6 @@ from PySide6.QtWidgets import (
     QCheckBox,
     QColorDialog,
     QComboBox,
-    QFileDialog,
     QGroupBox,
     QHBoxLayout,
     QHeaderView,
@@ -231,18 +230,16 @@ class SizeTable(QTableWidget):
             self.sizes_changed.emit()
 
     def _on_browse(self, size: int) -> None:
-        from icoforge.utils.theme import get_theme_manager
-        from icoforge.utils.window_theme import apply_theme_to_dialog
+        from chodzkos_gui_kit.qt.dialogs import open_file
 
-        dlg = QFileDialog(
+        path = open_file(
             self.window(),
             self.tr("Źródło dla %1x%2").replace("%1", str(size)).replace("%2", str(size)),
+            "",
+            _DIALOG_FILTER,
         )
-        dlg.setOption(QFileDialog.Option.DontUseNativeDialog, True)
-        dlg.setNameFilter(_DIALOG_FILTER)
-        apply_theme_to_dialog(dlg, get_theme_manager())
-        if dlg.exec() and dlg.selectedFiles():
-            self._set_override(size, Path(dlg.selectedFiles()[0]))
+        if path:
+            self._set_override(size, Path(path))
 
     def _on_context_menu(self, pos: QPoint) -> None:
         row = self.rowAt(pos.y())